% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d1}
\alias{d1}
\title{Example Dataset d1}
\format{
## `d1`
A data.frame with 1000 rows and 7 columns:
\describe{
  \item{id}{individuals}
  \item{year}{periods}
  \item{y}{dependent variable}
  \item{y1}{the first lag of y}
  \item{q}{threshold variable}
  \item{x}{regressor with threshold effects}
  \item{z}{regressor without threshold effects}
}
}
\source{
Simulated data with two thresholds
}
\usage{
d1
}
\description{
A simulated dataset for demonstrating the package
}
\keyword{datasets}
