% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{A simulation data used for examples}
\format{
A simulation data used for examples
\describe{
None

}
}
\usage{
data
}
\description{
A simulation data used for examples
}
\keyword{datasets}
