% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Growth_Inflation.R
\docType{data}
\name{Growth_Inflation}
\alias{Growth_Inflation}
\title{Example Dataset Growth_Inflation}
\format{
## `Growth_Inflation`
A data.frame with 814 rows and 15 columns:
\describe{
  \item{ncountry}{country id}
  \item{countryname}{country name}
  \item{code}{country code}
  \item{Period}{Period}
  \item{years}{years}
  \item{GDP per capita growth}{the difference of ln(GDP per capita)}
  \item{Inflation rate-semilog}{the semi-log of Inflation rate}
  \item{Transitional convergence}{the lag of ln(GDP per capita)}
  \item{Human capital}{Human capital}
  \item{Financial depth}{Financial depth}
  \item{Governance}{Governance}
  \item{Public infrastructure}{Public infrastructure}
  \item{Trade openness}{Trade openness}
  \item{Economic instability}{Economic instability}
  \item{Inflation rate}{Inflation rate}
}
}
\source{
https://doi.org/10.1080/07474938.2019.1624401
}
\usage{
Growth_Inflation
}
\description{
A dataset for economic growth of 74 countries from 1961 to 2015 (five-year average).
}
\keyword{datasets}
