% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPM.R
\name{DPML}
\alias{DPML}
\title{The dynamic panel linear model with fixed effects}
\usage{
DPML(
  y,
  y1 = NULL,
  x = NULL,
  w = NULL,
  var_u = NULL,
  tt,
  nn,
  assumption = 1,
  time_trend = FALSE,
  time_fix_effects = FALSE,
  restart = FALSE,
  x1 = NULL,
  delty0 = NULL,
  Only_b = FALSE
)
}
\arguments{
\item{y}{the dependent variable; vector type input.}

\item{y1}{the lag dependent variable; vector type input; By default y1 is NULL,
and then y1 will be computed by y automatically.}

\item{x}{the independent variable; matrix type input.}

\item{w}{the variance ratio; By default is NULL; It must be greater than 1, and
only works when assumption is 2.}

\item{var_u}{the option of variance of error term; By default is NULL; It must be
greater than 0; When meet relevant ERROR, please change the var_u.}

\item{tt}{the length of time period.}

\item{nn}{the number of individuals.}

\item{assumption}{the option of assumption; By default assumption is 1, and it can be 2;
More details see Hsiao (2002).}

\item{time_trend}{the time trend; By default it is FALSE.}

\item{time_fix_effects}{the time fixed effects; By default it is FALSE.}

\item{restart}{the option of iterations; By default restart is FALSE,
if encounters iteration failure, please set restart as TRUE.}

\item{x1}{the initial values of independent variable; matrix type input.
By default x1 is NULL, and thus x1 will be computed by x automatically.}

\item{delty0}{the option of delta_y; By default delty0 is NULL; Pleas do not change delty0.}

\item{Only_b}{the option of initial equation;By default Only_b is FALSE, and if Only_b is TRUE, initial delta y will be a constant C.
More details please see Hsiao (2002) and Ramirez-Rondan (2020).}
}
\value{
A List of estimate results.
}
\description{
The dynamic panel linear model with fixed effects
}
\section{Functions}{
\itemize{
\item \code{DPML()}: This is a dynamic panel linear model with fixed effects, which
allows time trend term or time fixed effects.

}}
\examples{
data("data", package = "DPTM")
y <- data$data_test_linear$y
q <- data$data_test_linear$q
x <- as.matrix(data$data_test_linear$x)
z <- as.matrix(data$data_test_linear$z)
tt <- data$data_test_linear$tt
nn <- data$data_test_linear$nn
xx <- cbind(x,z)
m1 <- DPML(y=y,x=xx,tt=tt,nn=nn,assumption = 1)
m1$Coefs
}
\references{
Ramirez-Rondan, N. R. (2020). Maximum likelihood estimation
of dynamic panel threshold models. Econometric Reviews, 39(3), 260-276.

Hsiao, C., Pesaran, M. H., & Tahmiscioglu, A. K. (2002).
Maximum likelihood estimation of fixed effects dynamic panel data models covering short time periods. Journal of econometrics, 109(1), 107-150.
}
\author{
Hujie Bai
}
